Essentially working
Now, the search is done on page load, assuming that the title can be
found within the page.  If there are results, they are passed to the
page action for display in the popup, and the action is shown.

some basic styling with bootstrap, need to add an icon.

The initial search is quite slow, not sure where the delay is.